Update 'Nomor Rekening' validation rule to require numeric values in EditRekening, Mitra, and User models

+++ app/Nova/Actions/EditRekening.php
@@ -67,7 +67,7 @@ class EditRekening extends Action
                 ->displayUsingLabels()
                 ->rules('required'),
             Text::make('Nomor Rekening', 'rekening')
-                ->rules('required', 'integer'),
+                ->rules('required', 'numeric'),
 
         ];
     }
